requirements.txt
Django==4.1.6
django-cors-headers==3.13.0
Pillow==9.4.0
qrcode==7.4.1